+++ CHANGES.md
@@ -3,6 +3,7 @@ dev
 
 - add `libexecinfo-dev` to Alpine dev packages as it is used
   by multicore OCaml (@avsm)
+- retrieve cygsympathy script from upstream master (@dra27)
 
 v7.1.0 2021-02-25 Cambridge
 ---------------------------